Add PrimeField::mul_assign.
Introduce a new method `mul_add_assign` and `mul_add` for computing
a * b + c using a single modular reduction. Benchmarks (and very bland test)
are attached can be run with:

 cargo bench ark_ff::mul_add

It's still not clear to me if this implementation is really worth addition:
if we remove the black_box() calls from the benchmark ark_ff::mul_add/raw,
we get roughly the same performance as mul_add.