Add katex support in documentation.
This commit adds latex support within rustdoc for arkworks algebra.
In order to add latex code in textmode, use delimiters `$` and `$`.
In order to add latex code in displaymode, use delimiters `$$` and `$$`.
To compile the documentation, use `cargo rustdoc`.

This is achieved via the option `--html-in-header` set via
- the environment variable `RUSTDOCFLAGS`, or
- the variable `build.rustdocflags` of `.cargo/config.toml`
I opted for the latter making it a default compilation flag to use with
rustdoc.

IMPORTANT: for some reason (independent from us) `cargo rustdoc` fails
on rust nightly and this is due to a failure concerning `cfg-if` and
`rand-chacha`. I did not investigate the reason why this happened.